fix(arch): move contracts package from interfaces to domain layer
The contracts package was moved to the interfaces layer in a previous PR,
but this violated the architecture layer dependencies:
- infrastructure cannot import from interfaces
- composition cannot import from interfaces

This commit moves the contracts to the domain layer, which is accessible
by all layers. The interfaces layer now re-exports from domain.contracts
for backward compatibility.

Changes:
- Create domain/contracts/ with Gold layer Pandera schemas
- Update composition/factories/pipeline_factories.py to import from domain
- Update infrastructure/schemas/gold.py backward-compat shim
- Update interfaces/contracts/ to re-export from domain.contracts
- Add publications.py to LOC exemptions (374 LOC)
- Export contracts module from domain.__init__.py
