+++ src/bioetl/domain/contracts/gold/__init__.py
@@ -1,20 +1,27 @@
-"""Gold layer Pandera schemas for data validation.
+"""Gold layer data contracts organized by provider.
 
-DEPRECATED: This module re-exports schemas from bioetl.domain.contracts for
-backward compatibility. New code should import from bioetl.domain.contracts.
+This package contains Pandera DataFrameModel schemas for Gold layer validation,
+organized by data provider for easy navigation.
 
-Example:
-    # Preferred (new code):
-    from bioetl.domain.contracts import ChEMBLActivityGoldSchema
+Submodules:
+    chembl: ChEMBL bioactivity database schemas
+    pubchem: PubChem compound schemas
+    uniprot: UniProt protein database schemas
+    publications: Cross-provider publication schemas (PubMed, CrossRef, OpenAlex, SemanticScholar)
 
-    # Deprecated (legacy code, still works):
-    from bioetl.infrastructure.schemas.gold import ChEMBLActivityGoldSchema
+Example usage:
+    >>> from bioetl.domain.contracts.gold import chembl
+    >>> df = pd.read_parquet("data/gold/chembl_activity/")
+    >>> chembl.ChEMBLActivityGoldSchema.validate(df)
+
+    >>> from bioetl.domain.contracts.gold.publications import PubMedPublicationGoldSchema
+    >>> PubMedPublicationGoldSchema.validate(pubmed_df)
 """
 
 from __future__ import annotations
 
-# Re-export all schemas from contracts for backward compatibility
-from bioetl.domain.contracts.gold import (
+# Import all schemas for flat namespace access
+from bioetl.domain.contracts.gold.chembl import (
     ChEMBLActivityGoldSchema,
     ChEMBLAssayGoldSchema,
     ChEMBLAssayParametersGoldSchema,
@@ -27,18 +34,20 @@ from bioetl.domain.contracts.gold import (
     ChEMBLProteinClassGoldSchema,
     ChEMBLTargetComponentGoldSchema,
     ChEMBLTargetGoldSchema,
+)
+from bioetl.domain.contracts.gold.pubchem import PubChemCompoundGoldSchema
+from bioetl.domain.contracts.gold.publications import (
     CrossRefPublicationGoldSchema,
     OpenAlexPublicationGoldSchema,
-    PubChemCompoundGoldSchema,
     PubMedPublicationGoldSchema,
     SemanticScholarPublicationGoldSchema,
+)
+from bioetl.domain.contracts.gold.uniprot import (
     UniProtIDMappingGoldSchema,
     UniProtProteinGoldSchema,
 )
-from bioetl.domain.contracts.gold._base import DATE_REGEX
 
 __all__ = [
-    "DATE_REGEX",
     "ChEMBLActivityGoldSchema",
     "ChEMBLAssayGoldSchema",
     "ChEMBLAssayParametersGoldSchema",
