+++ src/bioetl/interfaces/contracts/gold/pubchem.py
@@ -1,5 +1,8 @@
 """PubChem Gold layer data contracts.
 
+DEPRECATED: This module re-exports schemas from bioetl.domain.contracts.gold.pubchem for
+backward compatibility. New code should import from bioetl.domain.contracts.gold.pubchem.
+
 Contains Pandera DataFrameModel schemas for PubChem entities in the Gold layer:
 - Compound: Chemical structures and identifiers from PubChem
 
@@ -11,41 +14,7 @@ Int→Float coercion note:
 
 from __future__ import annotations
 
-import pandera.pandas as pa
-from pandera.typing import Series
-
-
-class PubChemCompoundGoldSchema(pa.DataFrameModel):
-    """Schema for PubChem Compound in Gold layer.
-
-    Aligned with domain/entities/pubchem.py (PubchemMolecule domain entity)
-    and application/pipelines/pubchem/transformer.py (PubChemCompoundTransformer).
-    """
-
-    entity_id: Series[str] = pa.Field(nullable=False)
-    cid: Series[str] = pa.Field(nullable=False)  # Domain entity uses str for cid
-    molecular_formula: Series[str] = pa.Field(nullable=True)
-    molecular_weight: Series[float] = pa.Field(
-        nullable=True, coerce=True
-    )  # Transformed to float by transformer
-    canonical_smiles: Series[str] = pa.Field(nullable=True)
-    isomeric_smiles: Series[str] = pa.Field(nullable=True)
-    inchi: Series[str] = pa.Field(nullable=True)
-    inchikey: Series[str] = pa.Field(nullable=True)
-    iupac_name: Series[str] = pa.Field(nullable=True)
-    content_hash: Series[str] = pa.Field(nullable=False)
-
-    # Metadata
-    run_id: Series[str] = pa.Field(nullable=False, alias="_run_id")
-    run_type: Series[str] = pa.Field(nullable=False, alias="_run_type")
-    source_batch_id: Series[str] = pa.Field(nullable=True, alias="_source_batch_id")
-    ingestion_ts: Series[str] = pa.Field(nullable=False, alias="_ingestion_ts")
-    index: Series[int] = pa.Field(nullable=False, alias="_index")
-
-    class Config:
-        """Pandera configuration for strict schema validation."""
-
-        strict = True
-
+# Re-export all schemas from domain.contracts for backward compatibility
+from bioetl.domain.contracts.gold.pubchem import PubChemCompoundGoldSchema
 
 __all__ = ["PubChemCompoundGoldSchema"]
