+++ src/bioetl/interfaces/contracts/gold/publications.py
@@ -1,5 +1,8 @@
 """Publication Gold layer data contracts.
 
+DEPRECATED: This module re-exports schemas from bioetl.domain.contracts.gold.publications
+for backward compatibility. New code should import from bioetl.domain.contracts.gold.publications.
+
 Contains Pandera DataFrameModel schemas for cross-provider publication entities
 in the Gold layer:
 - PubMed: Publication metadata with MEDLINE-specific fields
@@ -15,356 +18,13 @@ Int→Float coercion note:
 
 from __future__ import annotations
 
-import pandera.pandas as pa
-from pandera.typing import Series
-
-from bioetl.interfaces.contracts.gold._base import DATE_REGEX
-
-
-class PubMedPublicationGoldSchema(pa.DataFrameModel):
-    """Schema for PubMed Publication in Gold layer.
-
-    Includes PubMed-specific fields for forensic retention and detailed analysis.
-    See: https://dtd.nlm.nih.gov/ncbi/pubmed/out/pubmed_230101.dtd
-    """
-
-    entity_id: Series[str] = pa.Field(nullable=False)
-    content_hash: Series[str] = pa.Field(nullable=False)
-    pmid: Series[str] = pa.Field(nullable=False)
-    doi: Series[str] = pa.Field(nullable=True)
-    pmc_id: Series[str] = pa.Field(nullable=True)
-    title: Series[str] = pa.Field(nullable=True)
-    abstract: Series[str] = pa.Field(nullable=True)
-    abstract_structured: Series[bool] = pa.Field(
-        nullable=True
-    )  # Whether abstract has NLM sections
-    vernacular_title: Series[str] = pa.Field(
-        nullable=True
-    )  # Original non-English title
-
-    # Journal information
-    journal: Series[str] = pa.Field(nullable=True)
-    journal_abbrev: Series[str] = pa.Field(nullable=True)
-    # PubMed-specific journal fields (forensic retention)
-    journal_title: Series[str] = pa.Field(nullable=True)  # Full journal name (PubMed)
-    journal_iso_abbrev: Series[str] = pa.Field(nullable=True)  # ISO abbreviation
-    journal_issn_type: Series[str] = pa.Field(
-        nullable=True
-    )  # ISSN type: Print/Electronic/Linking
-    issn: Series[str] = pa.Field(nullable=True)
-    nlm_unique_id: Series[str] = pa.Field(nullable=True)  # NLM catalog ID
-    volume: Series[str] = pa.Field(nullable=True)
-    issue: Series[str] = pa.Field(nullable=True)
-
-    # Page information
-    pages: Series[str] = pa.Field(nullable=True)  # Legacy: medline_pgn format
-    medline_pgn: Series[str] = pa.Field(nullable=True)  # Original PubMed pagination
-    first_page: Series[str] = pa.Field(nullable=True)  # Unified: parsed from pages
-    last_page: Series[str] = pa.Field(nullable=True)  # Unified: parsed from pages
-
-    # Authors
-    authors: Series[str] = pa.Field(nullable=True)  # JSON-serialized list
-
-    # Date fields
-    pub_date: Series[str] = pa.Field(nullable=True)
-    pub_month: Series[float] = pa.Field(nullable=True, coerce=True)  # Month (1-12)
-    pub_day: Series[float] = pa.Field(nullable=True, coerce=True)  # Day (1-31)
-    publication_date: Series[str] = pa.Field(
-        nullable=True, str_matches=DATE_REGEX
-    )  # Unified: YYYY-MM-DD
-    year: Series[float] = pa.Field(nullable=True, coerce=True)
-    publication_year: Series[float] = pa.Field(
-        nullable=True, coerce=True
-    )  # Legacy alias
-    accepted_date: Series[str] = pa.Field(nullable=True, str_matches=DATE_REGEX)
-    received_date: Series[str] = pa.Field(nullable=True, str_matches=DATE_REGEX)
-    revised_date: Series[str] = pa.Field(nullable=True, str_matches=DATE_REGEX)
-    epub_date: Series[str] = pa.Field(nullable=True, str_matches=DATE_REGEX)
-    # MEDLINE-specific dates
-    date_completed: Series[str] = pa.Field(
-        nullable=True, str_matches=DATE_REGEX
-    )  # MEDLINE processing completion
-    date_revised: Series[str] = pa.Field(
-        nullable=True, str_matches=DATE_REGEX
-    )  # Record revision date (MEDLINE)
-
-    # Publication status and types
-    publication_status: Series[str] = pa.Field(
-        nullable=True
-    )  # ppublish/epublish/aheadofprint
-    publication_type_list: Series[str] = pa.Field(
-        nullable=True
-    )  # JSON array of pub types
-    publication_types: Series[object] = pa.Field(nullable=True)  # list[str]
-
-    # Classification
-    keywords: Series[object] = pa.Field(nullable=True)  # list[str]
-    mesh_terms: Series[object] = pa.Field(nullable=True)  # list[str]
-    citation_subset: Series[str] = pa.Field(
-        nullable=True
-    )  # Citation subset codes (e.g., 'AIM')
-    language: Series[str] = pa.Field(nullable=True)
-    country: Series[str] = pa.Field(nullable=True)
-
-    # Counts (denormalized for query efficiency)
-    author_count: Series[float] = pa.Field(nullable=True, coerce=True)
-    mesh_heading_count: Series[float] = pa.Field(nullable=True, coerce=True)
-    keyword_count: Series[float] = pa.Field(nullable=True, coerce=True)
-    grant_count: Series[float] = pa.Field(nullable=True, coerce=True)
-    reference_count: Series[float] = pa.Field(nullable=True, coerce=True)
-    chemical_count: Series[float] = pa.Field(nullable=True, coerce=True)
-
-    # Source tracking
-    source: Series[str] = pa.Field(nullable=True)
-
-    # Lookup metadata
-    # _lookup_method: "direct" | "doi" | "pmid" | "title_fallback" | "unknown"
-    # _original_id: Original identifier used for lookup
-    lookup_method: Series[str] = pa.Field(nullable=True, alias="_lookup_method")
-    original_id: Series[str] = pa.Field(nullable=True, alias="_original_id")
-
-    # DQ fields
-    dq_warn: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_warn")
-    dq_error: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_error")
-
-    # Metadata
-    run_id: Series[str] = pa.Field(nullable=False, alias="_run_id")
-    run_type: Series[str] = pa.Field(nullable=False, alias="_run_type")
-    source_batch_id: Series[str] = pa.Field(nullable=True, alias="_source_batch_id")
-    ingestion_ts: Series[str] = pa.Field(nullable=False, alias="_ingestion_ts")
-    index: Series[int] = pa.Field(nullable=False, alias="_index")
-
-    class Config:
-        """Pandera configuration for strict schema validation."""
-
-        strict = True
-
-
-class CrossRefPublicationGoldSchema(pa.DataFrameModel):
-    """Schema for CrossRef Publication in Gold layer.
-
-    Used for enriching publication records with CrossRef metadata via DOI resolution.
-    """
-
-    # System fields
-    entity_id: Series[str] = pa.Field(nullable=False)
-    content_hash: Series[str] = pa.Field(nullable=False)
-
-    # Primary identifier
-    # doi: Digital Object Identifier (lowercase, without "https://doi.org/") - Primary key
-    # Note: CrossRef uses DOI as primary key; pmid/pmc_id come from PubMed, not CrossRef
-    doi: Series[str] = pa.Field(nullable=False)
-
-    # Cross-reference IDs for linking publications across providers
-    # pmid: PubMed ID (numeric string: "12345678") - null for CrossRef native records
-    pmid: Series[str] = pa.Field(nullable=True)
-    # pmc_id: PubMed Central ID (format: "PMC1234567") - null for CrossRef native records
-    pmc_id: Series[str] = pa.Field(nullable=True)
-
-    # Core fields
-    title: Series[str] = pa.Field(nullable=True)
-    abstract: Series[str] = pa.Field(nullable=True)
-    authors: Series[str] = pa.Field(nullable=True)  # JSON-serialized list
-    journal: Series[str] = pa.Field(nullable=True)
-    issn: Series[object] = pa.Field(nullable=True)  # list[str]
-    publisher: Series[str] = pa.Field(nullable=True)
-    volume: Series[str] = pa.Field(nullable=True)
-    issue: Series[str] = pa.Field(nullable=True)
-    first_page: Series[str] = pa.Field(nullable=True)
-    last_page: Series[str] = pa.Field(nullable=True)
-
-    # Date fields
-    year: Series[float] = pa.Field(nullable=True, ge=1900, le=2100, coerce=True)
-    publication_date: Series[str] = pa.Field(
-        nullable=True, str_matches=DATE_REGEX
-    )  # Unified: YYYY-MM-DD
-    published_print: Series[str] = pa.Field(
-        nullable=True, str_matches=DATE_REGEX
-    )  # Legacy: provider-specific
-    published_online: Series[str] = pa.Field(
-        nullable=True, str_matches=DATE_REGEX
-    )  # Legacy: provider-specific
-
-    # Metadata
-    doc_type: Series[str] = pa.Field(nullable=True)
-    citation_count: Series[float] = pa.Field(nullable=True, ge=0, coerce=True)
-    reference_count: Series[float] = pa.Field(nullable=True, ge=0, coerce=True)
-    language: Series[str] = pa.Field(nullable=True)
-    license_url: Series[str] = pa.Field(nullable=True)
-    subjects: Series[object] = pa.Field(nullable=True)  # list[str]
-    source: Series[str] = pa.Field(nullable=True)
-
-    # Lookup metadata
-    # _lookup_method: "direct" | "doi" | "pmid" | "title_fallback" | "unknown"
-    # _original_id: Original identifier used for lookup
-    lookup_method: Series[str] = pa.Field(nullable=True, alias="_lookup_method")
-    original_id: Series[str] = pa.Field(nullable=True, alias="_original_id")
-
-    # DQ fields
-    dq_warn: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_warn")
-    dq_error: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_error")
-
-    # Lineage metadata
-    run_id: Series[str] = pa.Field(nullable=False, alias="_run_id")
-    run_type: Series[str] = pa.Field(nullable=False, alias="_run_type")
-    source_batch_id: Series[str] = pa.Field(nullable=True, alias="_source_batch_id")
-    ingestion_ts: Series[str] = pa.Field(nullable=False, alias="_ingestion_ts")
-    index: Series[int] = pa.Field(nullable=False, alias="_index")
-
-    class Config:
-        """Pandera configuration for strict schema validation."""
-
-        strict = True
-
-
-class OpenAlexPublicationGoldSchema(pa.DataFrameModel):
-    """Schema for OpenAlex Publication in Gold layer.
-
-    Used for batch DOI resolution with title fallback via OpenAlex Works API.
-    """
-
-    # System fields
-    entity_id: Series[str] = pa.Field(nullable=False)
-    content_hash: Series[str] = pa.Field(nullable=False)
-
-    # Primary key
-    openalex_id: Series[str] = pa.Field(nullable=False)
-
-    # Cross-reference IDs for linking publications across providers
-    # doi: Digital Object Identifier (lowercase, without "https://doi.org/")
-    doi: Series[str] = pa.Field(nullable=True)
-    # pmid: PubMed ID (numeric string: "12345678")
-    pmid: Series[str] = pa.Field(nullable=True)
-    # pmc_id: PubMed Central ID (format: "PMC1234567")
-    pmc_id: Series[str] = pa.Field(nullable=True)
-    title: Series[str] = pa.Field(nullable=True)
-    abstract: Series[str] = pa.Field(nullable=True)
-    authors: Series[str] = pa.Field(nullable=True)  # JSON-serialized list
-    concepts: Series[object] = pa.Field(nullable=True)  # list[str]
-
-    # Journal info
-    journal: Series[str] = pa.Field(nullable=True)
-    issn: Series[str] = pa.Field(nullable=True)
-    publisher: Series[str] = pa.Field(nullable=True)
-
-    # Page fields (nullable - OpenAlex API doesn't typically provide page information)
-    # Added for schema consistency across all publication providers
-    first_page: Series[str] = pa.Field(nullable=True)
-    last_page: Series[str] = pa.Field(nullable=True)
-
-    # Date fields
-    year: Series[float] = pa.Field(nullable=True, ge=1500, le=2100, coerce=True)
-    publication_date: Series[str] = pa.Field(nullable=True, str_matches=DATE_REGEX)
-
-    # Metadata
-    doc_type: Series[str] = pa.Field(nullable=False)
-    is_oa: Series[bool] = pa.Field(nullable=True)
-    oa_status: Series[str] = pa.Field(nullable=True)
-    # OpenAlex source field: cited_by_count
-    # Unified BioETL field: citation_count (standardized across all providers)
-    citation_count: Series[float] = pa.Field(nullable=True, ge=0, coerce=True)
-    language: Series[str] = pa.Field(nullable=True)
-    source: Series[str] = pa.Field(nullable=False)
-
-    # Lookup metadata
-    # _lookup_method: "direct" | "doi" | "pmid" | "title_fallback" | "unknown"
-    # _original_id: Original identifier used for lookup
-    lookup_method: Series[str] = pa.Field(nullable=False, alias="_lookup_method")
-    original_id: Series[str] = pa.Field(nullable=True, alias="_original_id")
-
-    # DQ fields
-    dq_warn: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_warn")
-    dq_error: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_error")
-
-    # Lineage metadata
-    run_id: Series[str] = pa.Field(nullable=False, alias="_run_id")
-    run_type: Series[str] = pa.Field(nullable=False, alias="_run_type")
-    source_batch_id: Series[str] = pa.Field(nullable=True, alias="_source_batch_id")
-    ingestion_ts: Series[str] = pa.Field(nullable=False, alias="_ingestion_ts")
-    index: Series[int] = pa.Field(nullable=False, alias="_index")
-
-    class Config:
-        """Pandera configuration for strict schema validation."""
-
-        strict = True
-
-
-class SemanticScholarPublicationGoldSchema(pa.DataFrameModel):
-    """Schema for Semantic Scholar Publication in Gold layer.
-
-    Used for enriching publication records with Semantic Scholar metadata.
-    """
-
-    # System fields
-    entity_id: Series[str] = pa.Field(nullable=False)
-    content_hash: Series[str] = pa.Field(nullable=False)
-
-    # Primary key
-    paper_id: Series[str] = pa.Field(nullable=False)
-
-    # External IDs
-    doi: Series[str] = pa.Field(nullable=True)
-    pmid: Series[str] = pa.Field(nullable=True)
-    # Cross-reference IDs for linking publications across providers
-    # pmc_id: PubMed Central ID (format: "PMC1234567")
-    pmc_id: Series[str] = pa.Field(nullable=True)
-    arxiv_id: Series[str] = pa.Field(nullable=True)
-    corpus_id: Series[float] = pa.Field(nullable=True, coerce=True)  # int64
-
-    # Core fields
-    title: Series[str] = pa.Field(nullable=True)
-    abstract: Series[str] = pa.Field(nullable=True)
-    tldr: Series[str] = pa.Field(nullable=True)
-    year: Series[float] = pa.Field(nullable=True, coerce=True)  # int64
-    publication_date: Series[str] = pa.Field(nullable=True, str_matches=DATE_REGEX)
-
-    # Journal/Venue
-    journal: Series[str] = pa.Field(nullable=True)
-    volume: Series[str] = pa.Field(nullable=True)
-    pages: Series[str] = pa.Field(nullable=True)  # Legacy: "first-last" format
-    first_page: Series[str] = pa.Field(nullable=True)  # Unified: parsed from pages
-    last_page: Series[str] = pa.Field(nullable=True)  # Unified: parsed from pages
-    venue: Series[str] = pa.Field(nullable=True)
-
-    # Metrics
-    citation_count: Series[float] = pa.Field(nullable=True, ge=0, coerce=True)  # int64
-    reference_count: Series[float] = pa.Field(nullable=True, ge=0, coerce=True)  # int64
-
-    # Open Access
-    is_oa: Series[bool] = pa.Field(nullable=True)
-    open_access_url: Series[str] = pa.Field(nullable=True)
-    oa_status: Series[str] = pa.Field(nullable=True)
-
-    # Classification (JSON strings)
-    fields_of_study: Series[str] = pa.Field(nullable=True)
-    publication_types: Series[str] = pa.Field(nullable=True)
-    authors: Series[str] = pa.Field(nullable=True)
-
-    # Source tracking
-    source: Series[str] = pa.Field(nullable=True)
-
-    # Lookup metadata
-    # _lookup_method: "direct" | "doi" | "pmid" | "title_fallback" | "unknown"
-    # _original_id: Original identifier used for lookup
-    lookup_method: Series[str] = pa.Field(nullable=True, alias="_lookup_method")
-    original_id: Series[str] = pa.Field(nullable=True, alias="_original_id")
-
-    # DQ fields
-    dq_warn: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_warn")
-    dq_error: Series[bool] = pa.Field(nullable=False, default=False, alias="_dq_error")
-
-    # Lineage metadata
-    run_id: Series[str] = pa.Field(nullable=False, alias="_run_id")
-    run_type: Series[str] = pa.Field(nullable=False, alias="_run_type")
-    source_batch_id: Series[str] = pa.Field(nullable=True, alias="_source_batch_id")
-    ingestion_ts: Series[str] = pa.Field(nullable=False, alias="_ingestion_ts")
-    index: Series[int] = pa.Field(nullable=False, alias="_index")
-
-    class Config:
-        """Pandera configuration for strict schema validation."""
-
-        strict = True
-
+# Re-export all schemas from domain.contracts for backward compatibility
+from bioetl.domain.contracts.gold.publications import (
+    CrossRefPublicationGoldSchema,
+    OpenAlexPublicationGoldSchema,
+    PubMedPublicationGoldSchema,
+    SemanticScholarPublicationGoldSchema,
+)
 
 __all__ = [
     "CrossRefPublicationGoldSchema",
